fix: split up slow tests
